Ein kreis um den Spieler herum mit einem Punkt symbolisiert nun die Waffe aus der die Schüsse geschossen werden. Die schüsse kommen nun nicht mehr aus der Person sondern aus der Hand

+++ src/main/java/com/dt/minigame/model/Bullet.java
@@ -18,8 +18,8 @@ public class Bullet {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private int id;
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private double angle;
 
     @Override
